Update tests for arena
