+++ src/typescript/sdk/package.json
@@ -109,9 +109,10 @@
     "pre-commit:install": "pre-commit install -c ../../../cfg/pre-commit-config.yaml",
     "pre-commit:run": "pre-commit run --all-files -c ../../../cfg/pre-commit-config.yaml",
     "prepublishOnly": "pnpm clean && pnpm i && pnpm check && pnpm build:publish",
-    "test": "pnpm run test:sdk:arena && pnpm run test:sdk:parallel && pnpm run test:sdk:sequential && pnpm run test:unit",
+    "test": "pnpm run test:sdk:arena && pnpm run test:sdk:arena-ws && pnpm run test:sdk:parallel && pnpm run test:sdk:sequential && pnpm run test:unit",
     "test:e2e": "pnpm run test:sdk:parallel && pnpm run test:sdk:sequential",
-    "test:sdk:arena": "pnpm jest tests/e2e/arena --runInBand",
+    "test:sdk:arena": "pnpm jest tests/e2e/arena/general --runInBand",
+    "test:sdk:arena-ws": "pnpm jest tests/e2e/arena/ws --runInBand",
     "test:sdk:parallel": "pnpm jest tests/e2e --testPathIgnorePatterns=\"(arena/|broker/)\"",
     "test:sdk:sequential": "pnpm jest tests/e2e/broker --runInBand",
     "test:unit": "pnpm jest tests/unit",
